update crop and preprocess
