+++ .gitignore
@@ -1,0 +1,1 @@
+openwillis/measures/video/__pycache__/
